requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
requests==2.32.4
urllib3==2.4.0
Werkzeug==3.1.3
